I think i moved things around to implement MVC
