+++ src/db.ts
@@ -1,10 +1,10 @@
 import {DataSource} from 'typeorm'
-import {Usuario} from './entities/Usuario'
-import {Compras} from './entities/Compras'
-import {DetalladoCompras} from './entities/DetalladoCompras'
-import {Productos} from './entities/Productos'
-import {ProductosEnPromocion} from './entities/ProductosEnPromocion'
-import {TipoPromocion} from './entities/TipoPromocion'
+import {Usuario} from './models/Usuario'
+import {Compras} from './models/Compras'
+import {DetalladoCompras} from './models/DetalladoCompras'
+import {Productos} from './models/Productos'
+import {ProductosEnPromocion} from './models/ProductosEnPromocion'
+import {TipoPromocion} from './models/TipoPromocion'
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
 
 export const AppDataSource= new DataSource({
