+++ src/models/DetalladoCompras.ts
