+++ src/models/ProductosEnPromocion.ts
