+++ src/models/TipoPromocion.ts
